Print NA to cf.stat_locus file if locus is not decisive for the branch (requested by Rob Lanfear). --cf-verbose (instead of --scf-part) to print cf.stat_locus and cf.stat_tree file.

+++ alignment/superalignment.h
@@ -251,7 +251,7 @@ public:
      @param quartet ID of four taxa
      @param[out] support number of sites supporting 12|34, 13|24 and 14|23
      */
-    virtual void computeQuartetSupports(IntVector &quartet, vector<size_t> &support);
+    virtual void computeQuartetSupports(IntVector &quartet, vector<int64_t> &support);
     
 	/**
 		@return unconstrained log-likelihood (without a tree)
